Move frdp components under server/frdp
