+++ server/frdp/frdp-authd/frdp-authd.c
