+++ server/frdp/frdp-krb-authd/frdp-krb-authd.c
