+++ server/frdp/frdp-sesmand/frdp-sesmand.c
