+++ server/frdp/frdp-session-agent/frdp-session-agent.c
