+++ server/frdp/frdpd/frdpd_auth.c
