+++ server/frdp/frdpd/frdpd_auth.h
