+++ server/frdp/frdpd/frdpd_pam.c
