+++ server/frdp/frdpd/frdpd_pam.h
