+++ server/frdp/ipc/frdp-ipc.h
